Cleaned up the code and comments, made SmartDashboard into DogLog for Jeff.
