+++ src/main/java/frc/robot/Robot.java
@@ -10,7 +10,7 @@ import edu.wpi.first.wpilibj.TimedRobot;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandScheduler;
 import frc.robot.commands.ZeroArm;
-import frc.robot.subsystems.NewArmSubsystem;
+import frc.robot.subsystems.ArmSubsystem;
 import frc.robot.util.LoggedTalonFX;
 
 /**
@@ -20,7 +20,7 @@ import frc.robot.util.LoggedTalonFX;
  */
 public class Robot extends TimedRobot {
   private Command m_autonomousCommand;
-  private ZeroArm zeroArm = new ZeroArm(NewArmSubsystem.getInstance());
+  private ZeroArm zeroArm = new ZeroArm(ArmSubsystem.getInstance());
 
   private final RobotContainer m_robotContainer;
 
@@ -66,15 +66,15 @@ public class Robot extends TimedRobot {
 
   /** This autonomous runs the autonomous command selected by your {@link RobotContainer} class. */
   @Override
-  // public void autonomousInit() {
-  //   RobotContainer.setAlliance();
-  //   m_autonomousCommand = m_robotContainer.getAutonomousCommand();
+  public void autonomousInit() {
+    RobotContainer.setAlliance();
+    m_autonomousCommand = m_robotContainer.getAutonomousCommand();
 
-  //   // schedule the autonomous command (example)
-  //   if (m_autonomousCommand != null) {
-  //     m_autonomousCommand.schedule();
-  //   }
-  // }
+    // schedule the autonomous command (example)
+    if (m_autonomousCommand != null) {
+      m_autonomousCommand.schedule();
+    }
+  }
 
   /** This function is called periodically during autonomous. */
   // @Override
