+++ src/main/java/frc/robot/commands/ArmToAngleCmd.java
@@ -2,50 +2,39 @@ package frc.robot.commands;
 
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.NewArmSubsystem;
+import frc.robot.subsystems.ArmSubsystem;
 import java.util.function.Supplier;
 
+import dev.doglog.DogLog;
+
 public class ArmToAngleCmd extends Command {
-  public ArmToAngleCmd(Supplier<Double> angle, NewArmSubsystem arm) {
+  public ArmToAngleCmd(Supplier<Double> angle, ArmSubsystem arm) {
     this.angle = angle;
     this.arm = arm;
     addRequirements(arm);
   }
 
-  private final NewArmSubsystem arm;
+  private final ArmSubsystem arm;
   private final Supplier<Double> angle;
-  private double endToleranceDegrees = 5;
-  private boolean returnToRestIfInterrupted = true;
+
 
   @Override
   public void initialize() {
-    // SmartDashboard.putBoolean("RUNNNININGINNIG", true);
   }
 
   @Override
   public void execute() {
-    SmartDashboard.putNumber("RUNNNINING Angle to target ", angle.get());
+    DogLog.log("Running to target angle: ", angle.get());
     arm.setPosition(angle.get());
   }
 
   @Override
   public boolean isFinished() {
-    // SmartDashboard.putBoolean("Is arm at target", arm.atTarget(endToleranceDegrees));
-    return false;
+    return arm.atTarget(5);
   }
 
   @Override
   public void end(boolean interrupted) {
-    // SmartDashboard.putBoolean("RUNNNININGINNIG BUNDT", false);
-    // if (returnToRestIfInterrupted && interrupted) {
-    //   // arm.rotateToRestPosition;
-    //   arm.resetPosition();
-    // }
-
   }
 
-  public ArmToAngleCmd withTolerance(double degrees) {
-    this.endToleranceDegrees = degrees;
-    return this;
-  }
 }
